backend/requirements.txt
fastapi
uvicorn[standard]
boto3
python-dotenv
SQLAlchemy
databases[sqlite]
pillow              # PIL.Image used in utils/*
python-multipart    # required for UploadFile in FastAPI
